# openapi-spec/openapi-unified.yaml
openapi: 3.0.0
info:
  title: E-Commerce Microservices API
  description: |
    Comprehensive API documentation for the E-Commerce Microservices platform.
    This unified specification documents all microservices APIs including:
    - Customer Service
    - Order Service
    - Inventory Service
    - Payment Service
    - Authentication Service
  version: '1.0.0'
  contact:
    name: E-Commerce API Support
    email: api-support@ecommerce.local
    url: https://ecommerce.local/support
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.staging.ecommerce.local
    description: Staging Environment
  - url: https://api.ecommerce.local
    description: Production Environment
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          enum:
            - RESOURCE_NOT_FOUND
            - VALIDATION_FAILED
            - INSUFFICIENT_STOCK
            - EVENT_PUBLISHING_ERROR
            - BUSINESS_ERROR
            - INTERNAL_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - AUTHENTICATION_FAILED
        message:
          type: string
          example: Resource not found
        timestamp:
          type: string
          format: date-time
        details:
          type: object
    PagedResponse:
      type: object
      required:
        - content
        - pageNumber
        - pageSize
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
          example: 0
        pageSize:
          type: integer
          format: int32
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          format: int32
          example: 5
...
    OrderResponse:
      type: object
      required:
        - id
        - customerId
        - status
        - totalAmount
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        customerId:
          type: integer
          format: int64
          example: 1
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - SHIPPED
            - DELIVERED
            - CANCELLED
            - FAILED
          example: PENDING
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
          example: 299.99
        paymentId:
          type: integer
          format: int64
        shippingAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          format: double
          example: 149.99
        subtotal:
          type: number
          format: double
          example: 299.98
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: integer
          format: int64
          example: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        shippingAddress:
          type: string
          example: '123 Main St, Anytown, USA'
    OrderItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2
...
    PaymentResponse:
      type: object
      required:
        - id
        - orderId
        - amount
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        orderId:
          type: integer
          format: int64
          example: 1
        amount:
          type: number
          format: double
          example: 299.99
        currency:
          type: string
          example: USD
        status:
          type: string
          enum:
            - PENDING
            - COMPLETED
            - FAILED
            - REFUNDED
          example: PENDING
        paymentMethod:
          type: string
          example: CREDIT_CARD
        transactionId:
          type: string
          example: txn_1234567890
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        RS256 access token issued by Keycloak (realm `ecommerce`). There is no login endpoint on this
        API: obtain a token from Keycloak's token endpoint, then send it as `Authorization: Bearer <token>`.
        See docs/KEYCLOAK_IDENTITY.md for roles and object-level rules.
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-indexed)
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        default: id
      example: id
security:
  - BearerAuth: []
paths:
...
  /api/orders:
    get:
      tags:
        - Order Management
      summary: Get all orders
      description: Retrieve all orders with pagination support
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        '200':
          description: Successfully retrieved orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrderResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Order Management
      summary: Create a new order
      description: Create a new order with resilience patterns (circuit breaker & retry)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable (circuit breaker open)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}:
    get:
      tags:
        - Order Management
      summary: Get order by ID
      description: Retrieve a specific order by their ID
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}/status:
    put:
      tags:
        - Order Management
      summary: Update order status
      description: Update the status of an existing order
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
          example: 1
        - name: status
          in: query
          required: true
          description: New order status
          schema:
            type: string
            enum:
              - PENDING
              - CONFIRMED
              - SHIPPED
              - DELIVERED
              - CANCELLED
              - FAILED
          example: CONFIRMED
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/payments/{id}:
    get:
      tags:
        - Payment Management
      summary: Get payment by ID
      description: Retrieve a specific payment by their ID
      parameters:
        - name: id
          in: path
          required: true
          description: Payment ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/payments/{id}/refund:
    post:
      tags:
        - Payment Management
      summary: Refund a payment
      description: Process a refund for an existing payment
      parameters:
        - name: id
          in: path
          required: true
          description: Payment ID
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Payment refunded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'